Use ServeContent instead.
http.ServeContent handles a lot of the details of HTTP responses nicely.
Since it also handles the Range header, make sure we don't use
compression when it is passed, so that valid data will be sent back.
Unfortunately http.ServeContent doesn't have a way to compress on the
fly, but in real usage, this won't make much of a difference since the file sizes
served by this blog will usually be small.